# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample Twitter server API documentation."
  version: "1.0.0"
  title: "Swagger Twitter"
host: "twitter.swagger.io"
basePath: "/v1"
...
schemes:
  - "http"
securityDefinitions:
  Bearer:
    description: Example value:- Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjU5MmQwMGJhNTJjYjJjM
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /tweets/createTweet:
    post:
      tags:
        - tweet
      security:
        - Bearer: []
      summary:
        Create a new tweet which has text and/or images. Also create a retweet
        of already created tweet
      operationId: createTweet
      parameters:
        - name: Create Tweet
          in: body
          required: true
          description: "Tweet details"
          schema:
            type: object
            properties:
              userId:
                type: string
                format: "ObjectId"
                description: "Required always. This is the id of the user who is posting new tweet or retweeting"
              tweetId:
                type: string
                format: "ObjectId"
                description: "Tweet id of the tweet which user is trying to retweet. Required when retweeting."
              userName:
                type: string
                description: "Required always. This is the username of the user who is posting new tweet or retweeting"
              userImageURL:
                type: string
                description: "Required always. This is the profile image url of the user who is posting new tweet or retweeting"
              imageURL:
                type: string
                description: "Optional. This is the URL of the image uploaded by the user along with the tweet"
              isRetweet:
                type: boolean
                description: "Required when user is retweeting. This field indicates that the user is retweeting a tweet"
              originalBody:
                type: string
                description: "Required when user is posting new tweet. This is field contains body of the tweet which the user is tweeting"
      responses:
        200:
          description: Tweet/Retweet created
        401:
          description: Invalid input
        404:
          description: Database not found
        501:
          description: Server not responding
  /tweets/addComment:
    post:
      tags:
        - tweet
      security:
        - Bearer: []
      summary: Add comment to a tweet
      description: Adds a comment to TweetId along with current time.
      operationId: addComment
      parameters:
        - name: comment
          in: body
          schema:
            type: object
            properties:
              tweetId:
                type: string
                format: ObjectId
              userId:
                type: string
                format: ObjectId
              userName:
                type: string
              imageURL:
                type: string
              body:
                type: string
      responses:
        200:
          description: Comment added successfully
        404:
          description: Delete comment not successful due to unavailable resources
        500:
          description: Internal server error
  /tweets/{tweetId}:
    delete:
      tags:
        - tweet
      security:
        - Bearer: []
      summary: Delete a tweet
      operationId: deleteTweet
      parameters:
        - name: tweetId
          in: path
          description: TweetId of the tweet which should be deleted
          required: true
          type: "string"
          format: "ObjectId"
      responses:
        200:
          description: Tweet deleted successfully
        404:
          description: Delete comment not successful due to unavailable resources
        500:
          description: Internal server error
...
  /fetchTweetByID/{TweetID} :
    get:
      tags:
        - "Search"
      security:
        - Bearer: []
      summary: "Fetch tweet using tweet ID"
      description: "Tweet ID is supplied and the tweet is returned"
      operationId: "fetchTweetByID"
      produces:
        - "application/json"
      parameters:
        - name: "tweetId"
          in: "path"
          description: "Id of the tweet to be fetched"
          required: true
          type: "number"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
...
  /user/likeTweet:
    post:
      tags:
        - tweet
      security:
        - Bearer: []
      summary: Users can like a tweet posted by another user.
      description:
        When user likes a tweet, store tweetid and userid in like table
        and increase the count of likes for the tweet.
      operationId: likeTweet
      parameters:
        - name: tweetid
          in: query
          description: TweetID of the tweet that needs to be liked
          required: true
          type: "string"
          format: "ObjectId"
        - name: userid
          in: query
          description: UserID/Username of the user that liked the tweet
          required: true
          type: "string"
          format: "ObjectId"
      responses:
        200:
          description: Like Tweet Successful
        204:
          description: Tweet not found
        400:
          description: User already liked the tweet
        401:
          description: Unauthorized
        404:
          description: Like Tweet not successful due to unavailable resources
        500:
          description: Internal server error
...